Move fromErr to utils
